+++ deployment/src/main/java/io/quarkiverse/ironjacamar/deployment/JCAProcessor.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.jca.deployment;
+package io.quarkiverse.ironjacamar.deployment;
 
 import java.util.List;
 import java.util.Set;
@@ -13,10 +13,10 @@ import org.jboss.jandex.IndexView;
 import org.jboss.jca.core.connectionmanager.pool.mcp.SemaphoreArrayListManagedConnectionPool;
 import org.jboss.jca.core.tx.jbossts.TransactionIntegrationImpl;
 
-import io.quarkiverse.jca.api.ResourceEndpoint;
-import io.quarkiverse.jca.runtime.JCARecorder;
-import io.quarkiverse.jca.runtime.connectionmanager.ConnectionManagerProducer;
-import io.quarkiverse.jca.spi.ResourceAdapterSupport;
+import io.quarkiverse.ironjacamar.ResourceAdapterSupport;
+import io.quarkiverse.ironjacamar.ResourceEndpoint;
+import io.quarkiverse.ironjacamar.runtime.JCARecorder;
+import io.quarkiverse.ironjacamar.runtime.connectionmanager.ConnectionManagerProducer;
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
 import io.quarkus.arc.deployment.UnremovableBeanBuildItem;
 import io.quarkus.arc.processor.DotNames;
