+++ runtime/src/main/java/io/quarkiverse/ironjacamar/runtime/ResourceAdapterShutdownListener.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.jca.runtime;
+package io.quarkiverse.ironjacamar.runtime;
 
 import java.util.Objects;
 import java.util.Set;
