+++ runtime/src/main/java/io/quarkiverse/ironjacamar/runtime/endpoint/DefaultMessageEndpointFactory.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.jca.runtime.endpoint;
+package io.quarkiverse.ironjacamar.runtime.endpoint;
 
 import java.lang.reflect.Method;
 
@@ -9,7 +9,7 @@ import jakarta.resource.spi.endpoint.MessageEndpoint;
 import jakarta.resource.spi.endpoint.MessageEndpointFactory;
 import jakarta.transaction.Transactional;
 
-import io.quarkiverse.jca.spi.ResourceAdapterSupport;
+import io.quarkiverse.ironjacamar.ResourceAdapterSupport;
 import io.quarkus.arc.Arc;
 
 public class DefaultMessageEndpointFactory implements MessageEndpointFactory {
